requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
gunicorn==21.2.0
httpx[http2]==0.24.1
pymongo[srv]